RESISTOME_SNPVERIFICATION updated; top 20 genes barchart added
