--- bin/AmrPlusPlus_SNP/config.ini
@@ -1,32 +1,0 @@
-[FULL_FILE_NAMES]
-; sample = PT221H2T0_GuelphTEAMR_S65_L002_dedup
-sample = S1_test
-
-[SOURCE_FILES]
-megares_fasta = data/megares_database_v3.00.fasta
-snp_info_fasta = data/SNPinfo.fasta
-count_matrix = data/sample_input/AMR_analytic_matrix.csv
-bam_input = data/sample_input/S1_test.bam
-
-[FOLDERS]
-main_output_folder = test_output/
-detailed_folder = detailed_output/
-
-[OUTPUT_FILES]
-count_matrix_final = AMR_analytic_matrix_with_SNP_confirmation.csv
-normal_type_output = Normal_Type_Genes.csv
-frameshift_type_output = Frameshift_Type_Genes.csv
-hypersusceptible_type_output = Hypersusceptible_Mutations_Type_Genes.csv
-suppressible_type_output = Suppressible_Frameshift_Type_Genes.csv
-intrinsic_type_output = Intrinsic_Resistance_Genes.csv
-
-[TEMP_FILES]
-temp_bam_sorted = temp/sorted.bam
-
-[SETTINGS]
-mt_and_wt = false
-detailed = false
-amrplusplus = false
-threads = 12
-debugging_mode = false
-
