--- bin/AmrPlusPlus_SNP/data/sample_input/AMR_analytic_matrix.csv
@@ -1,105 +1,0 @@
-gene_accession,S1_test,S2_test,S3_test
-MEG_2777|Drugs|MLS|23S_rRNA_methyltransferases|ERM37,0,0,8.0
-MEG_1920|Drugs|betalactams|Class_C_betalactamases|CMY,0,0,12.0
-MEG_4781|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_7312|Drugs|Elfamycins|EF-Tu_inhibition|TUFAB|RequiresSNPConfirmation,0,0,22.0
-MEG_1068|Drugs|Aminoglycosides|Aminoglycoside_O-phosphotransferases|APH3-PRIME,0,0,10.0
-MEG_5932|Drugs|Fluoroquinolones|Quinolone_resistance_protein_Qnr|QNRB,0,0,6.0
-MEG_2941|Drugs|Sulfonamides|Sulfonamide-resistant_dihydropteroate_synthases|FOLP|RequiresSNPConfirmation,0,0,16.0
-MEG_126|Drugs|Aminoglycosides|Aminoglycoside_N-acetyltransferases|AAC6-PRIME,0,0,6.0
-MEG_1543|Drugs|betalactams|Class_A_betalactamases|CARB,0,0,12.0
-MEG_6861|Drugs|betalactams|Class_A_betalactamases|TEM,8.0,0,0
-MEG_2279|Drugs|betalactams|Class_A_betalactamases|CTX,10.0,0,0
-MEG_1131|Drugs|Rifampin|Rifampin_ADP-ribosyltransferase_Arr|ARR,0,0,4.0
-MEG_6552|Multi-compound|Drug_and_biocide_and_metal_resistance|Drug_and_biocide_and_metal_resistance_regulator|SOXS|RequiresSNPConfirmation,4.0,4.0,4.0
-MEG_66|Drugs|Aminoglycosides|Aminoglycoside_N-acetyltransferases|AAC3,0,0,12.0
-MEG_7023|Drugs|Tetracyclines|Tetracycline_resistance_MFS_efflux_pumps|TETA,0,0,14.0
-MEG_2969|Drugs|Sulfonamides|Sulfonamide-resistant_dihydropteroate_synthases|FOLP|RequiresSNPConfirmation,0,0,10.0
-MEG_2344|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,16.0
-MEG_3637|Drugs|betalactams|Class_B_betalactamases|LRA,0,0,16.0
-MEG_6754|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,10.0
-MEG_3899|Multi-compound|Drug_and_biocide_resistance|Drug_and_biocide_RND_efflux_pumps|MEXA,0,0,12.0
-MEG_3045|Drugs|betalactams|Class_C_betalactamases|FOX,0,0,12.0
-MEG_5050|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,14.0
-MEG_7317|Drugs|Elfamycins|EF-Tu_inhibition|TUFAB|RequiresSNPConfirmation,0,0,18.0
-MEG_2828|Drugs|MLS|23S_rRNA_methyltransferases|ERMF,10.0,8.0,0
-MEG_1535|Drugs|betalactams|Class_A_betalactamases|CARB,0,0,8.0
-MEG_2420|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,10.0
-MEG_5972|Drugs|Fluoroquinolones|Quinolone_resistance_protein_Qnr|QNRB,0,0,6.0
-MEG_6702|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,16.0
-MEG_1588|Drugs|Phenicol|Chloramphenicol_acetyltransferases|CATA,0,0,12.0
-MEG_6379|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,10.0
-MEG_2417|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,14.0
-MEG_4778|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_3074|Drugs|Fusidic_acid|Fusidic_acid_esterases|FUSH,0,0,16.0
-MEG_6272|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,10.0
-MEG_1255|Drugs|betalactams|Class_B_betalactamases|BLAB,0,0,10.0
-MEG_5979|Drugs|Fluoroquinolones|Quinolone_resistance_protein_Qnr|QNRB,0,0,8.0
-MEG_5179|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_7316|Drugs|Elfamycins|EF-Tu_inhibition|TUFAB|RequiresSNPConfirmation,0,0,22.0
-MEG_6932|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,16.0
-MEG_447|Drugs|betalactams|Class_C_betalactamases|ACT,0,0,14.0
-MEG_4080|Multi-compound|Drug_and_biocide_resistance|Drug_and_biocide_RND_efflux_pumps|MTRC,16.0,0,20.0
-MEG_3362|Drugs|betalactams|Class_B_betalactamases|IMP,0,0,6.0
-MEG_1923|Drugs|betalactams|Class_C_betalactamases|CMY,0,0,12.0
-MEG_6733|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,10.0
-MEG_5062|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_4078|Drugs|Multi-drug_resistance|Multi-drug_RND_efflux_regulator|MTRAD,0,0,16.0
-MEG_1557|Drugs|Phenicol|Chloramphenicol_acetyltransferases|CAT,14.0,0,14.0
-MEG_284|Drugs|Aminoglycosides|Aminoglycoside_N-acetyltransferases|AAC6-PRIME,4.0,0,0
-MEG_6420|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,8.0
-MEG_4052|Drugs|MLS|Macrolide_phosphotransferases|MPHO,10.0,0,0
-MEG_2293|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,8.0
-MEG_7754|Drugs|betalactams|Class_B_betalactamases|VIM,0,0,6.0
-MEG_6003|Drugs|Fluoroquinolones|Quinolone_resistance_protein_Qnr|QNRB,0,0,6.0
-MEG_452|Drugs|betalactams|Class_C_betalactamases|ACT,0,0,14.0
-MEG_6471|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,8.0
-MEG_1579|Drugs|Phenicol|Chloramphenicol_acetyltransferases|CATA,0,0,8.0
-MEG_336|Drugs|Aminoglycosides|Aminoglycoside_N-acetyltransferases|AAC6-PRIME,6.0,0,0
-MEG_5066|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,8.0
-MEG_4760|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,12.0
-MEG_2105|Drugs|betalactams|Class_B_betalactamases|CPHA,0,0,10.0
-MEG_6083|Drugs|Multi-drug_resistance|Multi-drug_MFS_efflux_pumps|ROSA,0,0,14.0
-MEG_4750|Drugs|betalactams|Class_D_betalactamases|OXA,14.0,0,0
-MEG_7331|Drugs|Glycopeptides|VanA-type_resistance_protein|VAN,0,0,4.0
-MEG_3567|Drugs|betalactams|Class_A_betalactamases|LEN,0,0,6.0
-MEG_4872|Drugs|betalactams|Class_D_betalactamases|OXA,6.0,0,0
-MEG_6447|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,4.0
-MEG_5099|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_6698|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,12.0
-MEG_5395|Drugs|Fluoroquinolones|Fluoroquinolone-resistant_DNA_topoisomerases|PAREF|RequiresSNPConfirmation,0,0,28.0
-MEG_3009|Drugs|Fosfomycin|Fosfomycin_thiol_transferases|FOSB,0,0,4.0
-MEG_1836|Drugs|betalactams|Class_C_betalactamases|CMY,0,0,16.0
-MEG_2272|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,12.0
-MEG_6250|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,12.0
-MEG_414|Multi-compound|Drug_and_biocide_resistance|Drug_and_biocide_RND_efflux_regulator|ACRR|RequiresSNPConfirmation,0,0,8.0
-MEG_2943|Drugs|Sulfonamides|Sulfonamide-resistant_dihydropteroate_synthases|FOLP|RequiresSNPConfirmation,6.0,6.0,6.0
-MEG_4026|Drugs|MLS|Macrolide_phosphotransferases|MPH,0,0,20.0
-MEG_2847|Drugs|MLS|23S_rRNA_methyltransferases|ERMT,0,0,10.0
-MEG_3037|Drugs|betalactams|Class_C_betalactamases|FOX,0,0,14.0
-MEG_6006|Drugs|Fluoroquinolones|Quinolone_resistance_protein_Qnr|QNRB,0,0,6.0
-MEG_5199|Drugs|betalactams|Class_D_betalactamases|OXA,8.0,0,0
-MEG_2247|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,10.0
-MEG_6792|Drugs|betalactams|Class_A_betalactamases|TEM,0,0,10.0
-MEG_2421|Drugs|betalactams|Class_A_betalactamases|CTX,0,0,18.0
-MEG_4758|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,16.0
-MEG_4179|Drugs|Metronidazole|nim_nitroimidazole_reductase|NIMA,8.0,0,0
-MEG_6433|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,10.0
-MEG_6492|Drugs|betalactams|Class_A_betalactamases|SHV,0,0,12.0
-MEG_6775|Drugs|betalactams|Class_A_betalactamases|TEM,10.0,0,0
-MEG_1091|Drugs|Aminoglycosides|Aminoglycoside_O-phosphotransferases|APH6,0,0,10.0
-MEG_2842|Drugs|MLS|23S_rRNA_methyltransferases|ERMQ,0,0,10.0
-MEG_4611|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_4772|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_7728|Drugs|betalactams|Class_B_betalactamases|VIM,0,0,8.0
-MEG_1251|Drugs|betalactams|Class_B_betalactamases|BLAB,8.0,0,12.0
-MEG_4980|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_1083|Drugs|Aminoglycosides|Aminoglycoside_O-phosphotransferases|APH4,0,0,16.0
-MEG_1322|Drugs|betalactams|Class_A_betalactamases|BLAZ,0,0,8.0
-MEG_1773|Drugs|Phenicol|Phenicol_resistance_MFS_efflux_pumps|CMLA,0,0,8.0
-MEG_4986|Drugs|betalactams|Class_D_betalactamases|OXA,0,0,10.0
-MEG_7309|Drugs|Elfamycins|EF-Tu_inhibition|TUFAB|RequiresSNPConfirmation,20.0,0,16.0
-MEG_2579|Drugs|Trimethoprim|Dihydrofolate_reductase|DFRG,0,0,4.0
-MEG_6068|Drugs|Aminoglycosides|16S_rRNA_methyltransferases|RMTD,8.0,6.0,10.0
-MEG_445|Drugs|betalactams|Class_C_betalactamases|ACT,18.0,0,0
-MEG_7629|Drugs|MLS|Streptogramin_A_O-acetyltransferase|VATE,0,0,12.0
