+++ conf/modules.config
@@ -445,7 +445,23 @@ process {
             pattern: '*.log'
         ]
     }
-    
+
+    withName: RESISTOME_RESULTS {
+        publishDir = [
+            path: { "${params.outdir}/resistome_results/" },
+            mode: params.publish_dir_mode,
+            pattern: '*.csv'
+        ]
+    }
+   
+    withName: RESISTOME_SNPRESULTS {
+        publishDir = [
+            path: { "${params.outdir}/resistome_results/" },
+            mode: params.publish_dir_mode,
+            pattern: '*.csv'
+        ]
+    } 
+
     withName: SOURMASH_SKETCH {
         publishDir = [
             path: { "${params.outdir}/sourmash_sketch/" },
