+++ environment.yml
@@ -4,3 +4,5 @@ seaborn==0.12.2
 matplotlib==3.5.1
 scikit-learn==1.2.2
 multiqc==1.14
+pysam==0.22.0
+biopython=1.81
