+++ nextflow.config
@@ -162,6 +162,8 @@ params {
     amr_index_name                        = "./amrplusplus_ref/megares_database_v3.00.fasta"
     amr_fasta                             = "s3://zymo-filesystem/home/nsharp/amrplusplus_ref/megares_database_v3.00.fasta"
     amr_annotation                        = "s3://zymo-filesystem/home/nsharp/amrplusplus_ref/megares_annotations_v3.00.csv"
+    amr_SNP_config                        = "s3://zymo-filesystem/home/nsharp/amrplusplus_ref/config.ini"
+    amr_SNP_verification_database         = "s3://zymo-filesystem/home/nsharp/amrplusplus_ref/SNP_verification/*{.csv,.fasta}"
     threshold                             = 0.80 
 }
